requirements.txt
Flask
python-dotenv
openai
flask_cors
pycryptodome